Fix numeric types and possible overflows
There was a serious bug allowing to pass a value such as 3294967295 (u32) into the `/pvp` command, which is converted into -1000000001 (i32). As a result, the winner (ironically) will become a total bankrupt!

+++ src/config.rs
@@ -10,7 +10,7 @@ use crate::help;
 #[derive(Clone)]
 pub struct AppConfig {
     pub features: FeatureToggles,
-    pub top_limit: u32,
+    pub top_limit: u16,
 }
 
 #[derive(Clone)]
